Fix sending multipleids
