+++ src/index.html
@@ -56,6 +56,7 @@
         </div>
         <div class="modal-body d-flex flex-column align-items-center">
           <div class="d-flex flex-column align-items-center">
+            <div class="id-album"></div>
             <iframe id="album-player" width="250" height="300" allowfullscreen="" allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture"></iframe>
               <div class="d-flex">
                 <button id="add-comments" type="button" class="btn add-no-click info mb-0 mt-2 me-3">
